Problem fixed: Ultrasonic sensor reports invalid distance if it is higher than 127.
Reason:
Robot returns distance (range) which is always a positive number. It is a 2-byte number which is unsigned. Therefore, all the values smaller than 128 cm (2^7dec = 1000 0000 bit) are interpreted correctly (in 1 byte), higher numbers will be wrong, because Java interprets lower byte as signed. So we have to reinterpret it as if it is unsigned byte.

Solution:
We have to use int than short data type to store the result because short is again signed and can not fit the whole range of unsigned short (2-byte number).

Note: int is usually faster than short and occupies the same space if used as a single variable, because the Java natively is working with ints (and longs if it is 64bit Java).
http://stackoverflow.com/questions/2380696/java-short-integer-long-performance

(In Intel CPUs, for instance, handling 16-bit integers requires special prefix before each instruction, which slows down instructions by 1 cycle.)

Reinterpret signed byte as unsigned number.

How:
e.g. The distance between sensor and obstacle is 233cm
                              1110 1001    -23     (what we get as byte)
1111 1111 1111 1111 1111 1111 1110 1001    -23dec  (what we get as int)
0000 0000 0000 0000 0000 0000 1111 1111    255dec  (mask 0xFF, binary AND)
---------------------------------------
0000 0000 0000 0000 0000 0000 1110 1001    233dec  (correct int)