Implemented Speed Controller
Implement class MotorSpeed containing the speed values for the left and right wheel, respectively.

Introduced an Interface SpeedController for getting access to the current speed requested by the user. Implements it by setting the speed of the robot through touch positions on the screen.

The communication had to be put in another thread in order for the main thread to be able to process user inputs. Used an ScheduledExecutorService for this.

Changed the minimum SDK from 16 to 15